Added distance to prevent unintended reordering
Added a small distance delay on the sortable to prevent users from unintended
triggering of the reordering. (Happens all the time i'm using slickgrid).
i.e. the users tries to sort the column but fails since his mouse was still
on the move a little bit and 1px is enough to trigger reordeing here.
Please try, it still feels very responsive.